backend/requirements.txt
Django == 5.0.6
djangorestframework == 3.15.1
pymongo == 4.7.2
django-cors-headers == 4.4.0
requests == 2.32.3